chore: update package version to 0.3.2 and dependencies
- Bump version from 0.3.1 to 0.3.2 in package-lock.json
- Update browserslist from 4.27.0 to 4.28.0
- Upgrade baseline-browser-mapping to 2.8.25
- Upgrade caniuse-lite to 1.0.30001754
- Upgrade electron-to-chromium to 1.5.249
- Upgrade node-releases to 2.0.27
- Update sass from 1.93.3 to 1.94.0

+++ internal/repository/migrations.go
@@ -253,8 +253,8 @@ var migrations = []Migration{
 			case "mysql":
 				queries = []string{
 					`CREATE TABLE IF NOT EXISTS refresh_tokens (
-						id INT AUTO_INCREMENT PRIMARY KEY,
-						user_id INT NOT NULL,
+						id BIGINT AUTO_INCREMENT PRIMARY KEY,
+						user_id BIGINT NOT NULL,
 						token VARCHAR(255) UNIQUE NOT NULL,
 						expires_at DATETIME NOT NULL,
 						created_at DATETIME NOT NULL,
